create helpers.js and move youtubeAPI search logic into it
