+++ routes/index.js
@@ -1,9 +1,7 @@
 const express = require('express');
 const router = express.Router();
 const roomController = require('../controllers/roomController');
-const youtubeKey = "AIzaSyCe0L76JiEKzBGrrw1ByQHJKLLMD_e3MHg"
-const request = require('request');
-const BASE_URL = "https://www.googleapis.com/youtube/v3/";
+const helpers = require('../helpers');
 
 router.get('/', (req, res) => {
   res.render("index");
@@ -16,24 +14,7 @@ router.post('/rooms/create', roomController.createRoom);
 router.get('/rooms/edit', (req, res) =>{res.render("edit");});
 router.get('/rooms/:id', roomController.singleRoom);
 router.post(`/search`, (req, res)=> {
-  let response = {
-    BASE_URL: BASE_URL,
-    method: 'GET',
-    type: 'search?',
-    youtubeKey: youtubeKey,
-    part: 'snippet',
-    maxResults: 25,
-    q: 'cat',
-    order: 'relevance'
-  }
-  let getUrl = function(res){
-    return `${res.BASE_URL}${res.type}part=${res.part}&q=$${res.q}&maxResults=${res.maxResults}&order=${res.order}&key=${res.youtubeKey}`
-  }
-  //
-  request(getUrl(response), function(error, response, body){
-    console.log(response.body);
-  })
-  console.log(getUrl(res));
+  helpers.youtubeAPIRequest('GET', 'search', 'snippet', 25, 'cat');
   res.redirect(`/rooms/${req.body.id}`);
 });
